+++ commfuncs.cpp
@@ -2164,6 +2164,12 @@ short int encodeFunction(string str)
     return COUNTSTR;
   else if(sUpper.compare("FIELDNAME")==0)
     return FIELDNAME;
+  else if(sUpper.compare("CONCAT")==0)
+    return CONCAT;
+  else if(sUpper.compare("CONCATCOL")==0)
+    return CONCATCOL;
+  else if(sUpper.compare("CALCOL")==0)
+    return CALCOL;
   else if(sUpper.compare("ZONECONVERT")==0)
     return ZONECONVERT;
   else if(sUpper.compare("RANDOM")==0)